+++ test/check_tag.c
@@ -35,13 +35,13 @@ CheckShortField(TIFF *tif, const ttag_t field, const uint16_t value)
 	uint16_t tmp = 123;
 
 	if (!TIFFGetField(tif, field, &tmp)) {
-		fprintf (stderr, "Problem fetching tag %lu.\n",
-			 (unsigned long) field);
+		fprintf (stderr, "Problem fetching tag %"PRIu32".\n",
+			 field);
 		return -1;
 	}
 	if (tmp != value) {
-		fprintf (stderr, "Wrong SHORT value fetched for tag %lu.\n",
-			 (unsigned long) field);
+		fprintf (stderr, "Wrong SHORT value fetched for tag %"PRIu32".\n",
+			 field);
 		return -1;
 	}
 
@@ -54,13 +54,13 @@ CheckShortPairedField(TIFF *tif, const ttag_t field, const uint16_t *values)
 	uint16_t tmp[2] = {123, 456 };
 
 	if (!TIFFGetField(tif, field, tmp, tmp + 1)) {
-		fprintf (stderr, "Problem fetching tag %lu.\n",
-			 (unsigned long) field);
+		fprintf (stderr, "Problem fetching tag %"PRIu32".\n",
+			 field);
 		return -1;
 	}
 	if (tmp[0] != values[0] || tmp[1] != values[1]) {
-		fprintf (stderr, "Wrong SHORT PAIR fetched for tag %lu.\n",
-			 (unsigned long) field);
+		fprintf (stderr, "Wrong SHORT PAIR fetched for tag %"PRIu32".\n",
+			 field);
 		return -1;
 	}
 
@@ -73,13 +73,13 @@ CheckLongField(TIFF *tif, const ttag_t field, const uint32_t value)
 	uint32_t tmp = 123;
 
 	if (!TIFFGetField(tif, field, &tmp)) {
-		fprintf (stderr, "Problem fetching tag %lu.\n",
-			 (unsigned long) field);
+		fprintf (stderr, "Problem fetching tag %"PRIu32".\n",
+			 field);
 		return -1;
 	}
 	if (tmp != value) {
-		fprintf (stderr, "Wrong LONG value fetched for tag %lu.\n",
-			 (unsigned long) field);
+		fprintf (stderr, "Wrong LONG value fetched for tag %"PRIu32".\n",
+			 field);
 		return -1;
 	}
